+++ src/router/account.ts
@@ -1,0 +1,11 @@
+import express from "express";
+import register from "./account/register";
+import signin from "./account/signin";
+import verify from "./account/verify";
+import resend from "./account/resend";
+const router = express.Router();
+router.use(register);
+router.use(signin);
+router.use(verify);
+router.use(resend);
+export default router;