+++ src/router/conversation.ts
@@ -1,5 +1,5 @@
 import express from "express";
-import threads from "./conversation/threads";
+import threads from "./conversation/thread";
 import comment from "./conversation/comment";
 import create from "./conversation/create";
 import checkexist from "./conversation/checkexist";