+++ src/router/logout.ts
@@ -6,7 +6,7 @@ import express from "express";
 const router = express.Router();
 router.get("/api/logout", (req, res) => {
   res.cookie("key", "none", {
-    expires: new Date(Date.now() + 5),
+    expires: new Date(Date.now() + 1),
     httpOnly: true,
     domain: process.env.domain,
   });